refactor(views): migrate songs to per-folder layout (commit 6/7)
Split src/views/songs.rs (1236 lines) into src/views/songs/{mod,update,view}.rs.
Songs is the only non-expansion view in the family — no parent/child rows,
so the update fn does NOT use impl_expansion_update! and the file shape
diverges slightly from the other migrations (no expansion field on the
Page struct, simpler ViewPage trait coverage, update signature is
(message, songs) instead of (message, total_items, parents)).

- mod.rs: types (Page, Message, Action, ViewData), trait impls
  (HasCommonAction via macro with all three NavigateAndExpand variants,
  ViewPage explicit), Default + new(), tests for column-visibility
  defaults
- update.rs: impl SongsPage { fn update, fn sort_mode_to_api_string }
  + the toggle-column test
- view.rs: impl SongsPage { fn view } + the per-mode column helpers
  (songs_stars_visible / songs_plays_visible / songs_genre_visible) +
  helper tests. Row rendering stays inline in the view closure (the
  songs view has no separate render_row method).

Macro consumers:
- define_view_columns! for SongsColumn / Visibility (9 columns).
- impl_has_common_action!(SongsAction { NavigateAndExpandAlbum,
  NavigateAndExpandArtist, NavigateAndExpandGenre }) — songs is the
  one view that navigates to all three siblings.

cargo build / clippy --all-targets -D warnings / nightly fmt --check /
test --workspace all clean (662 + 368 + 1 doctest).
