+++ src/views/songs/view.rs
@@ -1,105 +1,18 @@
-//! Songs Page Component
+//! Songs view — `impl SongsPage { fn view }`.
 //!
-//! Self-contained songs view with slot list navigation, search, and filtering.
-//! Uses message bubbling pattern to communicate global actions to root.
-
-use std::collections::HashMap;
+//! Rendering for the songs page, plus the per-mode column-visibility helpers
+//! (`songs_stars_visible`, `songs_plays_visible`, `songs_genre_visible`).
+//!
+//! Update/state logic lives in `update.rs`; types live in `mod.rs`.
 
 use iced::{
-    Alignment, Element, Length, Task,
-    widget::{Row, button, container, image},
+    Alignment, Element, Length,
+    widget::{Row, button, container},
 };
-use nokkvi_data::{backend::songs::SongUIViewData, utils::formatters};
+use nokkvi_data::utils::formatters;
 
-use crate::widgets::{self, SlotListPageState, view_header::SortMode};
-
-/// Songs page local state
-#[derive(Debug)]
-pub struct SongsPage {
-    pub common: SlotListPageState,
-    /// Per-column visibility toggles surfaced via the columns-cog dropdown.
-    pub column_visibility: SongsColumnVisibility,
-}
-
-/// Toggleable songs columns. Index/Art/Title+Artist are always shown; the
-/// dynamic 18% slot still auto-renders Date/Year when sorted by those modes;
-/// Genre lives in the album column slot via the dedicated Genre toggle. Stars
-/// and Plays are now dedicated columns.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SongsColumn {
-    Select,
-    Index,
-    Thumbnail,
-    Stars,
-    Album,
-    Duration,
-    Plays,
-    Love,
-    Genre,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SongsColumnVisibility {
-    pub select: bool,
-    pub index: bool,
-    pub thumbnail: bool,
-    pub stars: bool,
-    pub album: bool,
-    pub duration: bool,
-    pub plays: bool,
-    pub love: bool,
-    pub genre: bool,
-}
-
-impl Default for SongsColumnVisibility {
-    fn default() -> Self {
-        // Stars and Plays opt-in; Album/Duration/Love always-on today.
-        // Index/Thumbnail default on to match historical always-on rendering.
-        // Genre opt-in; auto-shows when sort = Genre regardless of toggle.
-        // Select opt-in — checkbox column for multi-selection.
-        Self {
-            select: false,
-            index: true,
-            thumbnail: true,
-            stars: false,
-            album: true,
-            duration: true,
-            plays: false,
-            love: true,
-            genre: false,
-        }
-    }
-}
-
-impl SongsColumnVisibility {
-    pub fn get(&self, col: SongsColumn) -> bool {
-        match col {
-            SongsColumn::Select => self.select,
-            SongsColumn::Index => self.index,
-            SongsColumn::Thumbnail => self.thumbnail,
-            SongsColumn::Stars => self.stars,
-            SongsColumn::Album => self.album,
-            SongsColumn::Duration => self.duration,
-            SongsColumn::Plays => self.plays,
-            SongsColumn::Love => self.love,
-            SongsColumn::Genre => self.genre,
-        }
-    }
-
-    pub fn set(&mut self, col: SongsColumn, value: bool) {
-        match col {
-            SongsColumn::Select => self.select = value,
-            SongsColumn::Index => self.index = value,
-            SongsColumn::Thumbnail => self.thumbnail = value,
-            SongsColumn::Stars => self.stars = value,
-            SongsColumn::Album => self.album = value,
-            SongsColumn::Duration => self.duration = value,
-            SongsColumn::Plays => self.plays = value,
-            SongsColumn::Love => self.love = value,
-            SongsColumn::Genre => self.genre = value,
-        }
-    }
-}
+use super::{SongsColumn, SongsMessage, SongsPage, SongsViewData};
+use crate::widgets::{self, view_header::SortMode};
 
 pub(crate) fn songs_stars_visible(sort: SortMode, user_visible: bool) -> bool {
     user_visible || matches!(sort, SortMode::Rating)
@@ -115,443 +28,9 @@ pub(crate) fn songs_genre_visible(sort: SortMode, user_visible: bool) -> bool {
     user_visible || matches!(sort, SortMode::Genre)
 }
 
-/// View data passed from root (read-only, borrows from app state to avoid allocations)
-pub struct SongsViewData<'a> {
-    pub songs: &'a [SongUIViewData],
-    pub album_art: &'a HashMap<String, image::Handle>, // album_id -> artwork
-    pub large_artwork: &'a HashMap<String, image::Handle>,
-    pub dominant_colors: &'a HashMap<String, iced::Color>,
-    pub window_width: f32,
-    pub window_height: f32,
-    pub scale_factor: f32,
-    pub modifiers: iced::keyboard::Modifiers,
-    pub total_song_count: usize,
-    pub loading: bool,
-    pub stable_viewport: bool,
-    /// True when this view is rendered inside the library browsing panel
-    /// (split-view, right pane). Used to suppress chrome that doesn't fit
-    /// the narrower pane — e.g. the "Center on Playing" header button.
-    pub in_browsing_panel: bool,
-    /// Whether the column-visibility checkbox dropdown is open (controlled
-    /// by `Nokkvi.open_menu`).
-    pub column_dropdown_open: bool,
-    /// Trigger bounds captured when the dropdown was opened.
-    pub column_dropdown_trigger_bounds: Option<iced::Rectangle>,
-    /// Borrowed reference to the root open-menu state, so per-row context
-    /// menus and the artwork-panel context menu can resolve their own
-    /// open/closed status.
-    pub open_menu: Option<&'a crate::app_message::OpenMenu>,
-}
-
-/// Messages for local song page interactions
-#[derive(Debug, Clone)]
-pub enum SongsMessage {
-    // Slot list navigation
-    SlotListNavigateUp,
-    SlotListNavigateDown,
-    SlotListSetOffset(usize, iced::keyboard::Modifiers),
-    SlotListScrollSeek(usize),
-    SlotListActivateCenter,
-    SlotListClickPlay(usize), // Click non-center to play directly (skip focus)
-    /// Click on a row's leading select checkbox — toggles `item_index` in
-    /// `selected_indices`. No play/highlight side effects.
-    SlotListSelectionToggle(usize),
-    /// Click on the tri-state "select all" header — fills selection with
-    /// every visible row, or clears if every visible row is already selected.
-    SlotListSelectAllToggle,
-    AddCenterToQueue, // Add centered song to queue (Shift+Q)
-
-    // Mouse click on star/heart (item_index, value)
-    ClickSetRating(usize, usize), // (item_index, rating 1-5)
-    ClickToggleStar(usize),       // item_index
-
-    // Context menu
-    ContextMenuAction(usize, crate::widgets::context_menu::LibraryContextEntry),
-
-    // View header
-    SortModeSelected(widgets::view_header::SortMode),
-    ToggleSortOrder,
-    SearchQueryChanged(String),
-
-    RefreshViewData,
-    CenterOnPlaying,
-    /// Sort dropdown's "Roulette" entry was selected — intercepted at the
-    /// root handler before the page's `update` runs.
-    Roulette,
-
-    // Data loading (moved from root Message enum)
-    SongsLoaded {
-        result: Result<Vec<SongUIViewData>, String>,
-        total_count: usize,
-        background: bool,
-        anchor_id: Option<String>,
-    },
-    SongsPageLoaded(Result<Vec<SongUIViewData>, String>, usize), // result, total_count (subsequent page)
-    /// Refresh artwork for a specific album (album_id)
-    RefreshArtwork(String),
-    /// Navigate to a view and apply an ID filter
-    NavigateAndFilter(crate::View, nokkvi_data::types::filter::LibraryFilter),
-    /// Navigate to Albums and auto-expand the album with this id (no filter set).
-    NavigateAndExpandAlbum(String),
-    /// Navigate to Artists and auto-expand the artist with this id (no filter set).
-    NavigateAndExpandArtist(String),
-    /// Navigate to Genres and auto-expand the genre with this id (no filter set).
-    NavigateAndExpandGenre(String),
-    ToggleColumnVisible(SongsColumn),
-    /// Column-dropdown open/close request — bubbled to root
-    /// `Message::SetOpenMenu`. Intercepted in `handle_songs` before the
-    /// page's `update` runs.
-    SetOpenMenu(Option<crate::app_message::OpenMenu>),
-    /// Artwork column drag handle event — intercepted at root, page never sees it.
-    ArtworkColumnDrag(crate::widgets::artwork_split_handle::DragEvent),
-}
-
-/// Actions that bubble up to root for global state mutation
-#[derive(Debug, Clone)]
-pub enum SongsAction {
-    PlaySongFromIndex(usize), // Play songs starting from index
-    AddBatchToQueue(nokkvi_data::types::batch::BatchPayload),
-    AddBatchToPlaylist(nokkvi_data::types::batch::BatchPayload),
-    ToggleStar(String, bool), // (song_id, star)
-    SetRating(String, usize), // (song_id, rating) - set absolute rating
-    LoadLargeArtwork(String), // album_id for artwork
-    LoadPage(usize),          // offset - trigger fetch of next page
-
-    SearchChanged(String),                                  // trigger reload
-    SortModeChanged(widgets::view_header::SortMode),        // trigger reload
-    SortOrderChanged(bool),                                 // trigger reload
-    RefreshViewData,                                        // trigger reload
-    PlayNextBatch(nokkvi_data::types::batch::BatchPayload), // Batch payload
-    PlayBatch(nokkvi_data::types::batch::BatchPayload),     // Play immediately
-    AddToPlaylist(String),                                  // song_id - add to playlist dialog
-    ShowInfo(Box<nokkvi_data::types::info_modal::InfoModalItem>), // Open info modal
-    ShowInFolder(String),        // relative path - open containing folder
-    RefreshArtwork(String),      // album_id - refresh artwork from server
-    FindSimilar(String, String), // (id, label) - Find similar to this song
-    TopSongs(String, String),    // (artist, label) - Find top songs by artist
-    CenterOnPlaying,
-    NavigateAndFilter(crate::View, nokkvi_data::types::filter::LibraryFilter), // Navigate to target view and filter
-    NavigateAndExpandAlbum(String), // album_id - navigate to Albums and auto-expand this album
-    NavigateAndExpandArtist(String), // artist_id - navigate to Artists and auto-expand this artist
-    NavigateAndExpandGenre(String), // genre_id - navigate to Genres and auto-expand this genre
-    ColumnVisibilityChanged(SongsColumn, bool),
-    None,
-}
-
-impl super::HasCommonAction for SongsAction {
-    fn as_common(&self) -> super::CommonViewAction {
-        match self {
-            Self::SearchChanged(_) => super::CommonViewAction::SearchChanged,
-            Self::SortModeChanged(m) => super::CommonViewAction::SortModeChanged(*m),
-            Self::SortOrderChanged(a) => super::CommonViewAction::SortOrderChanged(*a),
-            Self::RefreshViewData => super::CommonViewAction::RefreshViewData,
-            Self::CenterOnPlaying => super::CommonViewAction::CenterOnPlaying,
-            Self::NavigateAndFilter(v, f) => {
-                super::CommonViewAction::NavigateAndFilter(*v, f.clone())
-            }
-            Self::NavigateAndExpandAlbum(id) => {
-                super::CommonViewAction::NavigateAndExpandAlbum(id.clone())
-            }
-            Self::NavigateAndExpandArtist(id) => {
-                super::CommonViewAction::NavigateAndExpandArtist(id.clone())
-            }
-            Self::NavigateAndExpandGenre(id) => {
-                super::CommonViewAction::NavigateAndExpandGenre(id.clone())
-            }
-            Self::None => super::CommonViewAction::None,
-            _ => super::CommonViewAction::ViewSpecific,
-        }
-    }
-}
-
-impl Default for SongsPage {
-    fn default() -> Self {
-        Self {
-            common: SlotListPageState::new(
-                widgets::view_header::SortMode::RecentlyAdded,
-                false, // sort_ascending
-            ),
-            column_visibility: SongsColumnVisibility::default(),
-        }
-    }
-}
-
 impl SongsPage {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Update internal state and return actions for root
-    pub fn update(
-        &mut self,
-        message: SongsMessage,
-        songs: &[SongUIViewData],
-    ) -> (Task<SongsMessage>, SongsAction) {
-        let total_items = songs.len();
-
-        match message {
-            SongsMessage::SlotListNavigateUp => {
-                self.common.handle_navigate_up(total_items);
-                if let Some(center_idx) = self.common.get_center_item_index(total_items)
-                    && let Some(song) = songs.get(center_idx)
-                    && let Some(album_id) = &song.album_id
-                {
-                    return (
-                        Task::none(),
-                        SongsAction::LoadLargeArtwork(album_id.clone()),
-                    );
-                }
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::SlotListNavigateDown => {
-                self.common.handle_navigate_down(total_items);
-                if let Some(center_idx) = self.common.get_center_item_index(total_items)
-                    && let Some(song) = songs.get(center_idx)
-                    && let Some(album_id) = &song.album_id
-                {
-                    return (
-                        Task::none(),
-                        SongsAction::LoadLargeArtwork(album_id.clone()),
-                    );
-                }
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::SlotListSetOffset(offset, modifiers) => {
-                self.common
-                    .handle_slot_click(offset, total_items, modifiers);
-                if let Some(center_idx) = self.common.get_center_item_index(total_items)
-                    && let Some(song) = songs.get(center_idx)
-                    && let Some(album_id) = &song.album_id
-                {
-                    return (
-                        Task::none(),
-                        SongsAction::LoadLargeArtwork(album_id.clone()),
-                    );
-                }
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::SlotListScrollSeek(offset) => {
-                self.common.handle_set_offset(offset, total_items);
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::SlotListActivateCenter => {
-                if !self.common.slot_list.selected_indices.is_empty() {
-                    use nokkvi_data::types::batch::{BatchItem, BatchPayload};
-                    let payload = self
-                        .common
-                        .slot_list
-                        .selected_indices
-                        .iter()
-                        .filter_map(|&index| {
-                            songs.get(index).map(|s| {
-                                let item: nokkvi_data::types::song::Song = s.clone().into();
-                                BatchItem::Song(Box::new(item))
-                            })
-                        })
-                        .fold(BatchPayload::new(), |p, i| p.with_item(i));
-                    (Task::none(), SongsAction::PlayBatch(payload))
-                } else if let Some(center_idx) = self.common.get_center_item_index(total_items) {
-                    self.common.slot_list.flash_center();
-                    (Task::none(), SongsAction::PlaySongFromIndex(center_idx))
-                } else {
-                    (Task::none(), SongsAction::None)
-                }
-            }
-            SongsMessage::SlotListClickPlay(offset) => {
-                self.common.handle_set_offset(offset, total_items);
-                self.update(SongsMessage::SlotListActivateCenter, songs)
-            }
-            SongsMessage::SlotListSelectionToggle(offset) => {
-                self.common.handle_selection_toggle(offset, total_items);
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::SlotListSelectAllToggle => {
-                self.common.handle_select_all_toggle(total_items);
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::AddCenterToQueue => {
-                use nokkvi_data::types::batch::BatchItem;
-
-                let target_indices = self.common.get_queue_target_indices(total_items);
-
-                if target_indices.is_empty() {
-                    return (Task::none(), SongsAction::None);
-                }
-
-                let payload = super::expansion::build_batch_payload(target_indices, |i| {
-                    songs.get(i).map(|s| {
-                        let item: nokkvi_data::types::song::Song = s.clone().into();
-                        BatchItem::Song(Box::new(item))
-                    })
-                });
-
-                (Task::none(), SongsAction::AddBatchToQueue(payload))
-            }
-
-            SongsMessage::ClickSetRating(item_index, rating) => {
-                if let Some(song) = songs.get(item_index) {
-                    use nokkvi_data::utils::formatters::compute_rating_toggle;
-                    let current = song.rating.unwrap_or(0) as usize;
-                    let new_rating = compute_rating_toggle(current, rating);
-                    (
-                        Task::none(),
-                        SongsAction::SetRating(song.id.clone(), new_rating),
-                    )
-                } else {
-                    (Task::none(), SongsAction::None)
-                }
-            }
-            SongsMessage::ClickToggleStar(item_index) => {
-                if let Some(song) = songs.get(item_index) {
-                    return (
-                        Task::none(),
-                        SongsAction::ToggleStar(song.id.clone(), !song.is_starred),
-                    );
-                }
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::SortModeSelected(sort_mode) => {
-                use crate::widgets::SlotListPageAction;
-                match self.common.handle_sort_mode_selected(sort_mode) {
-                    SlotListPageAction::SortModeChanged(vt) => {
-                        (Task::none(), SongsAction::SortModeChanged(vt))
-                    }
-                    _ => (Task::none(), SongsAction::None),
-                }
-            }
-            SongsMessage::ToggleSortOrder => {
-                use crate::widgets::SlotListPageAction;
-                match self.common.handle_toggle_sort_order() {
-                    SlotListPageAction::SortOrderChanged(ascending) => {
-                        (Task::none(), SongsAction::SortOrderChanged(ascending))
-                    }
-                    _ => (Task::none(), SongsAction::None),
-                }
-            }
-            SongsMessage::SearchQueryChanged(query) => {
-                use crate::widgets::SlotListPageAction;
-                match self.common.handle_search_query_changed(query, total_items) {
-                    SlotListPageAction::SearchChanged(q) => {
-                        (Task::none(), SongsAction::SearchChanged(q))
-                    }
-                    _ => (Task::none(), SongsAction::None),
-                }
-            }
-
-            SongsMessage::ContextMenuAction(clicked_idx, entry) => {
-                use nokkvi_data::types::batch::BatchItem;
-
-                use crate::widgets::context_menu::LibraryContextEntry;
-
-                let target_indices = self.common.get_batch_target_indices(clicked_idx);
-
-                let payload = super::expansion::build_batch_payload(target_indices, |i| {
-                    songs.get(i).map(|s| {
-                        let item: nokkvi_data::types::song::Song = s.clone().into();
-                        BatchItem::Song(Box::new(item))
-                    })
-                });
-
-                if let Some(song) = songs.get(clicked_idx) {
-                    match entry {
-                        LibraryContextEntry::AddToQueue => {
-                            (Task::none(), SongsAction::AddBatchToQueue(payload))
-                        }
-                        LibraryContextEntry::AddToPlaylist => {
-                            // AddToPlaylist backend takes a Vec<String> of song IDs, or a batch?
-                            // We will emit AddBatchToPlaylist but for now, if Batch doesn't fit AddToPlaylist perfectly,
-                            // we can map payload -> IDs. Let's just pass payload.
-                            (Task::none(), SongsAction::AddBatchToPlaylist(payload))
-                        }
-                        LibraryContextEntry::GetInfo => {
-                            use nokkvi_data::types::info_modal::InfoModalItem;
-                            let item = InfoModalItem::from_song_view_data(song);
-                            (Task::none(), SongsAction::ShowInfo(Box::new(item)))
-                        }
-                        LibraryContextEntry::ShowInFolder => {
-                            (Task::none(), SongsAction::ShowInFolder(song.path.clone()))
-                        }
-                        LibraryContextEntry::FindSimilar => (
-                            Task::none(),
-                            SongsAction::FindSimilar(song.id.clone(), song.title.clone()),
-                        ),
-                        LibraryContextEntry::TopSongs => {
-                            let artist = &song.artist;
-                            if !artist.is_empty() {
-                                (
-                                    Task::none(),
-                                    SongsAction::TopSongs(
-                                        artist.clone(),
-                                        format!("Top Songs: {artist}"),
-                                    ),
-                                )
-                            } else {
-                                (Task::none(), SongsAction::None)
-                            }
-                        }
-                        LibraryContextEntry::Separator
-                        | LibraryContextEntry::ReplaceQueueWithAllFound
-                        | LibraryContextEntry::AddAllFoundToQueue
-                        | LibraryContextEntry::AddAllFoundToPlaylist => {
-                            (Task::none(), SongsAction::None)
-                        }
-                    }
-                } else {
-                    (Task::none(), SongsAction::None)
-                }
-            }
-
-            // Data loading messages (handled at root level, no action needed here)
-            SongsMessage::SongsLoaded { .. } | SongsMessage::SongsPageLoaded(_, _) => {
-                (Task::none(), SongsAction::None)
-            }
-            // Routed up to root in `handle_songs` before this match runs;
-            // arm exists only for exhaustiveness.
-            SongsMessage::SetOpenMenu(_) => (Task::none(), SongsAction::None),
-            SongsMessage::Roulette => (Task::none(), SongsAction::None),
-            SongsMessage::ArtworkColumnDrag(_) => {
-                // Intercepted at root before reaching this update; never reached.
-                (Task::none(), SongsAction::None)
-            }
-            SongsMessage::RefreshViewData => (Task::none(), SongsAction::RefreshViewData),
-            SongsMessage::RefreshArtwork(album_id) => {
-                (Task::none(), SongsAction::RefreshArtwork(album_id))
-            }
-            SongsMessage::CenterOnPlaying => (Task::none(), SongsAction::CenterOnPlaying),
-            SongsMessage::NavigateAndFilter(view, filter) => {
-                (Task::none(), SongsAction::NavigateAndFilter(view, filter))
-            }
-            SongsMessage::NavigateAndExpandAlbum(album_id) => {
-                (Task::none(), SongsAction::NavigateAndExpandAlbum(album_id))
-            }
-            SongsMessage::NavigateAndExpandArtist(artist_id) => (
-                Task::none(),
-                SongsAction::NavigateAndExpandArtist(artist_id),
-            ),
-            SongsMessage::NavigateAndExpandGenre(genre_id) => {
-                (Task::none(), SongsAction::NavigateAndExpandGenre(genre_id))
-            }
-            SongsMessage::ToggleColumnVisible(col) => {
-                let new_value = !self.column_visibility.get(col);
-                self.column_visibility.set(col, new_value);
-                (
-                    Task::none(),
-                    SongsAction::ColumnVisibilityChanged(col, new_value),
-                )
-            }
-        }
-    }
-
-    /// Convert SortMode to API string for ViewModel.
-    /// Thin shim — the unified mapping lives in `views/sort_api.rs`.
-    pub fn sort_mode_to_api_string(sort_mode: SortMode) -> &'static str {
-        super::sort_api::sort_mode_to_api_string(crate::View::Songs, sort_mode)
-    }
-
     /// Build the view
     pub fn view<'a>(&'a self, data: SongsViewData<'a>) -> Element<'a, SongsMessage> {
-        use crate::widgets::view_header::SortMode;
-
         let column_dropdown: Element<'a, SongsMessage> = {
             use crate::widgets::checkbox_dropdown::checkbox_dropdown;
             let items: Vec<(SongsColumn, &'static str, bool)> = vec![
@@ -1128,55 +607,9 @@ impl SongsPage {
     }
 }
 
-// ============================================================================
-// ViewPage trait implementation
-// ============================================================================
-
-use crate::app_message::Message;
-
-impl super::ViewPage for SongsPage {
-    fn common(&self) -> &SlotListPageState {
-        &self.common
-    }
-    fn common_mut(&mut self) -> &mut SlotListPageState {
-        &mut self.common
-    }
-
-    fn search_input_id(&self) -> &'static str {
-        super::SONGS_SEARCH_ID
-    }
-
-    fn sort_mode_options(&self) -> Option<&'static [SortMode]> {
-        Some(SortMode::SONG_OPTIONS)
-    }
-    fn sort_mode_selected_message(&self, mode: SortMode) -> Option<Message> {
-        Some(Message::Songs(SongsMessage::SortModeSelected(mode)))
-    }
-    fn toggle_sort_order_message(&self) -> Message {
-        Message::Songs(SongsMessage::ToggleSortOrder)
-    }
-
-    fn add_to_queue_message(&self) -> Option<Message> {
-        Some(Message::Songs(SongsMessage::AddCenterToQueue))
-    }
-    fn reload_message(&self) -> Option<Message> {
-        Some(Message::LoadSongs)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn songs_column_visibility_default_preserves_today_behavior() {
-        let v = SongsColumnVisibility::default();
-        assert!(!v.stars);
-        assert!(v.album);
-        assert!(v.duration);
-        assert!(!v.plays);
-        assert!(v.love);
-    }
 
     #[test]
     fn songs_stars_visible_auto_shows_on_rating_sort() {
@@ -1200,37 +633,5 @@ mod tests {
         assert!(songs_genre_visible(SortMode::Genre, true));
         assert!(!songs_genre_visible(SortMode::Title, false));
         assert!(songs_genre_visible(SortMode::Title, true));
-    }
-
-    #[test]
-    fn songs_column_visibility_default_keeps_genre_off() {
-        let v = SongsColumnVisibility::default();
-        assert!(!v.genre);
-    }
-
-    #[test]
-    fn songs_toggle_column_visible_flips_state_and_emits_action() {
-        let mut page = SongsPage::default();
-        let empty: Vec<SongUIViewData> = vec![];
-        let (_t, action) = page.update(
-            SongsMessage::ToggleColumnVisible(SongsColumn::Plays),
-            &empty,
-        );
-        assert!(page.column_visibility.plays);
-        assert!(matches!(
-            action,
-            SongsAction::ColumnVisibilityChanged(SongsColumn::Plays, true)
-        ));
-
-        // Genre default is off → toggle ON, message carries Genre+true.
-        let (_t, action) = page.update(
-            SongsMessage::ToggleColumnVisible(SongsColumn::Genre),
-            &empty,
-        );
-        assert!(page.column_visibility.genre);
-        assert!(matches!(
-            action,
-            SongsAction::ColumnVisibilityChanged(SongsColumn::Genre, true)
-        ));
     }
 }
